Consolidate affiliate data into a single source of truth in /affiliate folder
